+++ src/make-base.sh
@@ -37,16 +37,18 @@ echo "Active widgetset: $WIDGET"
 echo "Starting new Listaller build..."
 # Command line to build the sofware
 # Create necessary dirs
-mkdir -p ./build
-mkdir -p ./build/locale
+mkdir -p ../build
+mkdir -p ../build/locale
 
 echo "Compiling libinstaller library..."
-lazbuild -B --ws=nogui ./libs/libinstaller.lpr
-ln -s ./build/libinstaller.so.0.4.0 libinstaller.so.0.4
-ln -s libinstaller.so.0.4 libinstaller.so
+lazbuild -B --ws=nogui ../lib/libinstaller.lpr
+if [ ! -f "libinstaller.so" ]; then
+ ln -s ../build/libinstaller.so.0.4.0 libinstaller.so.0.4
+ ln -s libinstaller.so.0.4 libinstaller.so
+fi
 
 echo "Compiling Listaller daemon..."
-lazbuild -B --ws=nogui ./helper/listallerd.lpr
+lazbuild -B --ws=nogui ../helper/listallerd.lpr
 
 echo "Compiling command-line tool..."
 lazbuild -B --ws=nogui lipa.lpr
@@ -57,11 +59,11 @@ lazbuild -B --ws=nogui libuild.lpr
 
 #Compiling lanuage files
 echo "Generating language files..."
-for i in `find ./data/locale -name "*.po"`
+for i in `find ../data/locale -name "*.po"`
 do
 echo "Format $i"
 msgfmt -o `expr substr $i 1 $(( ${#i} - 3 ))`.mo $i
-mv `expr substr $i 1 $(( ${#i} - 3 ))`.mo ./build/locale/
+mv `expr substr $i 1 $(( ${#i} - 3 ))`.mo ../build/locale/
 done
 
 echo "Listaller build completed."
